frontend/requirements.txt
streamlit
streamlit_timeline
st_pages
httpx
httpx-sse
requests
plotly
pandas
python-dotenv
pyvis
langchain-neo4j
langchain-ollama
